git-epic-creator/services/gitlab_client_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gitlab-client-service"
version = "1.0.0"
description = "GitLab Client Service for Agentic AI Backlog System"
authors = [
    { name = "Vladyslav Avilov", email = "vladyslav.avilov@email.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "structlog",
    "httpx",
    "tenacity",
    "python-gitlab>=4.0.0",
    "openai>=1.0.0",
    "redis[hiredis]>=5.0.0",
    "authlib>=1.3.0",
    "shared @ file:../shared"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "httpx"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*"]


